+++ sky/unit/test/widget/align_test.dart
@@ -1,7 +1,7 @@
 import 'package:sky/widgets.dart';
 import 'package:test/test.dart';
 
-import 'build_utils.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Align smoke test', () {
