+++ sky/unit/test/widget/box_decoration_test.dart
@@ -2,7 +2,7 @@ import 'package:sky/theme/colors.dart' as colors;
 import 'package:sky/widgets.dart';
 import 'package:test/test.dart';
 
-import 'build_utils.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Circles can have uniform borders', () {
