+++ sky/unit/test/widget/date_picker_test.dart
@@ -1,7 +1,7 @@
 import 'package:sky/widgets.dart';
 import 'package:test/test.dart';
 
-import 'build_utils.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Can select a day', () {
