+++ sky/unit/test/widget/mimic_test.dart
@@ -1,7 +1,7 @@
 import 'package:sky/widgets.dart';
 import 'package:test/test.dart';
 
-import 'build_utils.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Mimic can track tree movements', () {
